requirements.txt
flask
pyelftools